==> code/01_Gathering_Tweets.ipynb <==
# Check (That Tweet) Yo Self 
## Prioritizing Tweets to Fact Check
###### Part 1: Gathering Tweets

For our project, we want to find a way to prioritize disaster related tweets to be fact checked in order to slow the spread of misinformation. To start, we'll gather tweets with a few keywords.

Import necessary libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import requests
import time
import warnings
import regex as re
import seaborn as sns


from sklearn.pipeline import Pipeline, FeatureUnion
from sklearn.preprocessing import StandardScaler, FunctionTransformer
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.metrics import confusion_matrix, accuracy_score, classification_report

from nltk.sentiment.vader import SentimentIntensityAnalyzer
warnings.filterwarnings('ignore')
np.random.seed(824)
from bs4 import BeautifulSoup 

# Import stopwords.
from nltk.corpus import stopwords # Import the stopword list
import nltk

%config InlineBackend.figure_format = 'retina'
%matplotlib inline

plt.style.use('fivethirtyeight')

from tweetscrape.profile_tweets import TweetScrapperProfile
from tweetscrape.search_tweets import TweetScrapperSearch
from tweetscrape.users_scrape import TweetScrapperUser

These are the keywords we will pull tweets about.

keywords = ['isolation',
            'pandemic',
            'COVID',
            'quarantine',
            'vaccine',
            'coronavirus',
            'lysol',
            'ingest',
            'inject',
            'disinfectant',
            'bleach']

The function below loops through the keyword and gathers 10,000 for each, then returns the final DataFrame with all tweets/keywords.

def get_tweets(words):
    to_concat = [] #to hold the dataframes of tweets
    for word in words: #for every keyword we want to search
        
        #gather the tweets and export to a csv
        tweet_scrapper = TweetScrapperSearch(search_all = word, search_till_date= '2020-04-26', num_tweets=10_000, tweet_dump_path=f'./data/initial_grab/{word}_tweets.csv', tweet_dump_format='csv')
        tweet_count, tweet_id, tweet_time, dump_path = tweet_scrapper.get_search_tweets()
        print('{0} tweets about {1} gathered till {2}'.format(tweet_count, word, tweet_time))
        
        #read the csv back in as a dataframe
        tweets = pd.read_csv(f'./data/initial_grab/{word}_tweets.csv')
        
        #add to list to concat later
        to_concat.append(tweets)
        
    #put all the tweets together and check the shape
    all_tweets = pd.concat(to_concat, axis = 0)
    print(f'Initial DataFrame was {all_tweets.shape}')
    
    #drop the duplicates, reset the index, and display the first five rows
    all_tweets = all_tweets.drop_duplicates()
    print(f'After dropping duplicate tweets, DataFrame is {all_tweets.shape}')
    
    #save as all_tweets csv file
    all_tweets.to_csv('./data/all_tweets.csv', index = False)
    return all_tweets

Since this gathering process was very slow, we included a lot of printouts so we knew it was still running.

get_tweets(keywords)

Reading out CSV of tweets back in to make sure it looks correct:

tweet = pd.read_csv('./data/all_tweets.csv')

tweet.head()
