==> code/02_Cleaning.ipynb <==
# Check (That Tweet) Yo Self 
## Prioritizing Tweets to Fact Check
###### Part 2: Data Cleaning
In this notebook, we'll clean the tweets we've gathered about Coronavirus.

Import necessary libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import requests
import time
import warnings
import regex as re
import seaborn as sns
import re

from sklearn.pipeline import Pipeline, FeatureUnion
from sklearn.preprocessing import StandardScaler, FunctionTransformer
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.metrics import confusion_matrix, accuracy_score, classification_report

from nltk.sentiment.vader import SentimentIntensityAnalyzer
warnings.filterwarnings('ignore')
np.random.seed(824)
from bs4 import BeautifulSoup 

# Import stopwords.
from nltk.corpus import stopwords # Import the stopword list
import nltk

%config InlineBackend.figure_format = 'retina'
%matplotlib inline

plt.style.use('fivethirtyeight')

### Read in collected dataframe

tweet = pd.read_csv('../data/all_tweets.csv')

### Drop duplicate tweets (a little over 100 tweets removed)

tweet.shape

tweet.head(2)

tweet = tweet.drop_duplicates(subset='id', keep="first")

tweet.shape

### Drop "re-tweeter" column: it comprises only null entires
### Drop "type" column: it comprises only "tweet" entries, no knowledge gain.

tweet.isnull().sum()

tweet['type'].value_counts()

tweet = tweet.drop(columns = ['re_tweeter', 'type'])

### Remove additional missing data

tweet.isnull().sum()

tweet = tweet.dropna()

### 1. Convert time column to readable text
### 2. Addition of day column to see time distribution of collected tweets

tweet.dtypes

tweet['time'] = pd.to_numeric(tweet['time'])

import time

def change_time(x):
    x = x / 1000
    return time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(x))

tweet['time'] = tweet['time'].apply(change_time)

def day_classification (string):
    if "2020-04-26" in string:
        return "Sunday"
    elif "2020-04-25" in string:
        return "Saturday"
    elif "2020-04-24" in string:
        return "Friday"
    elif "2020-04-23" in string:
        return "Thursday"

tweet['day'] = tweet['time'].map(day_classification)    

tweet['day'].value_counts()

### There are 5848 tweets that were collected that aren't within our timeframe of interest. Dropping these tweets from the data

tweet.isnull().sum()

tweet = tweet.dropna()

### Clean up text to be predominantly recognized as english by the sklearn library

#set english words
english = set(nltk.corpus.words.words())

def is_english(string):
    #split string up
    words = string.split()
    #start counter
    count = 0
    for word in words:
        #check if the word is in our english list
        if word.lower() not in english:
            count += 1
        #check if any characters are not in the alphabet
        elif word.lower().isalpha() == False:
            count += .33
    perc = round(count / len(words),4)
    #return the proportion of non-english words
    return perc

#apply to this function and create a new column
tweet['not_english'] = tweet['text'].apply(is_english)

tweet.head(2)

tweet.describe()

### Setting a criteria for amount of english within the tweet
- Example: 
- (tweet[tweet['not_english'] < .4]) 
- less than 40% of the tweet is not identified as english by sklearn (tweet is 60% english)

tweet[tweet['not_english'] < 0.5]

### using 50% english now, might change that later..

#only keep words that are more than 50% english
tweet = tweet[tweet['not_english'] < .5]

tweet.shape

tweet.isnull().sum()

### Utilize iterable hashtag and mantion items to add columns for counts of hashtags and mentions associated with each tweet

def to_list(string):
    #remove all unnecessary characters
    string = string.replace('[', '') 
    string = string.replace(']', '') 
    string = string.replace('\'', '')
    #split into list based on commas
    new_list = string.split(',')
    #remove whitespace
    new_list = [x.strip() for x in new_list]
    #return the list
    return new_list

tweet['hashtags'] = tweet['hashtags'].map(to_list)
tweet['mentions'] = tweet['mentions'].map(to_list)

def count_hash(count_list):
    count = 0
    for i in count_list:
        if '#' in i:
            count += 1
    return count

tweet['hashtag_count'] = tweet['hashtags'].map(count_hash)

tweet['hashtag_count'].value_counts(normalize = True)

def count_at(count_list):
    count = 0
    for i in count_list:
        if "@" in i:
            count += 1
    return count

tweet['mention_count'] = tweet['mentions'].map(count_at)

tweet['mention_count'].value_counts(normalize = True)

### Create columns for tweet word count and character count

def word_count(string):
    return len(string.split())

tweet['word_count'] = tweet['text'].map(word_count)

tweet['word_count'].value_counts(normalize = True)

def char_count(string):
    return len(string)

tweet['char_count'] = tweet['text'].map(char_count)

tweet['char_count'].value_counts(normalize = True)

tweet.head(2)

### Adding column for count of attached links to tweet

tweet['link_count'] = [1 if len(x) > 2 else 0 for x in tweet['links']]

### Sentiment Analysis

def sentiment_score(string):
    sia = SentimentIntensityAnalyzer()
    return sia.polarity_scores(string)['compound']

tweet['text_sentiment'] = tweet['text'].map(sentiment_score)

tweet.describe()

# Adding column of text with links removed

tweet['text_links_removed'] = [re.sub('(\(https:\/\/[^\s]+)|https:\/\/[^\s]+', '', row) for row in tweet['text']]


# sklearn's stopwords, extracted
sklearn_stopwords = list(CountVectorizer(stop_words = 'english').get_stop_words())
#Custom created list
#custom_stopwords = ['nosleep', 
#                    'scarystories',
#                    'quarantine',
#                    'virus', 
#                    'covid',
#                    'pandemic',
#                    'coronavirus',
#                    'corona',
#                    'don',
#                    've',
#                    'pt',
#                    'didn',
#                    'wasn',
#                    'll',
#                    'shouldn',
#                    'woo',
#                    'hadn'
#                    'notext',
#                   ]
# Personalized stopwords
#personal_stopwords = sklearn_stopwords + custom_stopwords
#print(personal_stopwords)

### Alter text data by removing HTML, non-letters, and stopwords in addition to tokenizing and converting text to all lowercase

def tweet_to_words(raw_tweet):
    
    # 1. Remove HTML.
    tweet_text = BeautifulSoup(raw_tweet).get_text()
    
    # 2. Remove non-letters.
    letters_only = re.sub("[^a-zA-Z]", " ", tweet_text)
    
    # 3. Convert to lower case, split into individual words.
    words = letters_only.lower().split()
    
    # 4. Convert personalized stopwords to set
    stops = set(stopwords.words('english')) #change this command to set(personal_stopwords) if you decide to customize

    # 5. Remove stopwords.
    meaningful_words = [w for w in words if w not in stops]
    
    # 6. Join the words back into one string separated by space, 
    # and return the result.
    return(" ".join(meaningful_words))

# Get the number of subreddits based on the dataframe size.
total_tweets = tweet.shape[0]
print(f'There are {total_tweets} tweets related to coronavirus.')

### Apply function to clean our tweet data

# Initialize an empty list to hold the clean posts.
clean_tweets = []

print("Cleaning and parsing twitter data...")

# Instantiate counter.
j = 0

# For every post in our training set...
for string in tweet['text_links_removed']:
    
    # Convert post to words, then append to clean_train_posts.
    clean_tweets.append(tweet_to_words(string))
    
    # If the index is divisible by 1000, print a message.
    if (j + 1) % 5000 == 0:
        print(f'Tweet {j + 1} of {total_tweets}.')
    
    j += 1  

### Add cleaned tweet text to dataframe 

tweet = tweet.assign(clean_text = clean_tweets)

### Creating word and character count columns for cleaned text

tweet['clean_word_count'] = tweet['clean_text'].map(word_count)

tweet['clean_char_count'] = tweet['clean_text'].map(char_count)

tweet.head()

tweet.describe()

tweet.to_csv('../data/cleaned_with_lysol.csv', index = False)
