==> code/04_Gathering_User_Data.ipynb <==
# Check (That Tweet) Yo Self 
## Prioritizing Tweets to Fact Check
###### Part 4: Gathering User Data

After exploring the tweets we gathered, we realized it would be hugely beneficial to also have information on the user like number of followers, number of tweets total, etc. In this notebook, we loop through our previously gathered tweets to add the user data. 

import pandas as pd
import numpy as np
from tweetscrape.users_scrape import TweetScrapperUser
import tweetscrape
import warnings
warnings.filterwarnings('ignore')
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

Importing our data

tweet = pd.read_csv('./data/cleaned_with_lysol.csv')

Quick check to see how the user info is returned from Twitter

ts = TweetScrapperUser("PulpNews")
user_info = ts.get_profile_info()

user_info['bio']

This function loops through each user to grab their profile information:

def get_user(df):
    #Start a counter to track progress
    count = 0
    
    #Create our new columns that will eventually hold the user data
    df['user_bio'] = 0
    df['user_location'] = 0
    df['user_url'] = 0
    df['user_tweets'] = 0
    df['user_following'] = 0
    df['user_followers'] = 0
    df['user_favorites'] = 0
    
    #Loop to gather user data and assign it to the correct column
    for user in df['author']:
        count += 1
        try:
            ts = TweetScrapperUser(user)
            user_info = ts.get_profile_info()
            df.loc[df['author'] == user, 'user_bio'] = user_info['bio']
            df.loc[df['author'] == user,'user_location'] = user_info['location']
            df.loc[df['author'] == user,'user_url'] = user_info['url']
            df.loc[df['author'] == user,'user_tweets'] = user_info['tweets']
            df.loc[df['author'] == user,'user_following'] = user_info['following']
            df.loc[df['author'] == user,'user_followers'] = user_info['followers']
            df.loc[df['author'] == user,'user_favorites'] = user_info['favorites']
            if count % 100 == 0:
                print(f'{count} users have been gathered')
                df.to_csv(f'./data/users/{count}_tweet_random.csv', index=False)
        except:
            pass
    
    #Return the final, filled in dataframe
    return df

### Pulling 40k random ones:

We already pulled around 80,000 tweets but going back to grab the user info ix very time consuming. To make it more efficient, we'll take a random set of 40,000 tweets to then grab user info. 

tweet_rand = tweet.sample(n = 40_000, random_state = 21).reset_index(drop=True)

tweet_rand

As mentioned above, this was extremely time consuming. While we had hopped for 40k, grabbing just over 30k took almost 12 hours. Hence why the cell below is unrun, after restarting the notebook we didn't want to wait another 12 hours for it to complete.

tweet_rand = get_user(tweet_rand)
tweet_rand.to_csv('./data/40k_random.csv', index = False)

Now that we have user data, we'll do some more EDA then get to modeling.
