==> code/05A_Linear_Regression_Phrases.ipynb <==
# Check (That Tweet) Yo Self 
## Prioritizing Tweets to Fact Check
###### Part 5A: Linear Model (Supervised)
Our goal is to predict the "reach" (retweet, favorite, and reply count) of a tweet. First, we will try linear

Imports

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import requests
import time
import warnings
import regex as re
import seaborn as sns
import re
import statistics

from sklearn.pipeline import Pipeline, FeatureUnion
from sklearn.preprocessing import StandardScaler, FunctionTransformer
from sklearn.linear_model import LogisticRegression, LinearRegression
from sklearn.linear_model import Ridge 
from sklearn.linear_model import Lasso 
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.metrics import confusion_matrix, accuracy_score, classification_report

from nltk.sentiment.vader import SentimentIntensityAnalyzer
warnings.filterwarnings('ignore')
np.random.seed(824)
from bs4 import BeautifulSoup 

# Import stopwords.
from nltk.corpus import stopwords # Import the stopword list
import nltk

from tweetscrape.users_scrape import TweetScrapperUser

%config InlineBackend.figure_format = 'retina'
%matplotlib inline

plt.style.use('fivethirtyeight')

Loading our data that now also contains user info:

tweet = pd.read_csv('../data/model_attempt_tweets.csv')

cvec = CountVectorizer(stop_words = 'english', min_df=1, max_df=0.25, ngram_range=(1, 3))




cvec_matrix = cvec.fit_transform(tweet['text'])

# Convert X_train into a DataFrame.

cvec_df = pd.DataFrame(cvec_matrix.toarray(),
                          columns=cvec.get_feature_names())
cvec_df

cvec_df.columns

These were some of the top phrases from vectorizing:

phrases = cvec_df[['hi er doc',
                   'er doc inject',
                   'hi er',
                   'doc inject consume',
                  'er doc',
                  'doc inject',
                  'consume disinfectants attempt',
                  'disinfectants attempt kill',
                  'attempt kill covid19',
                  'disinfectants attempt',
                  'attempt kill',
                  'doc',
                  'kill covid19',
                  'er',
                  'attempt',
                  'hi',
                  'hard lies',
                  'hard lies social',
                  'social isolation work']]

phrases

X = phrases

Engineering a new interaction term between "big feelings" and "follow ratio":

X['big_feelings_follow_ratio'] = tweet['big_feelings'] * tweet['follow_ratio']

y = tweet['target']

Trying another linear regression:

# Create train_test_split.
X_train, X_test, y_train, y_test = train_test_split(X,
                                                    y,
                                                   random_state = 42)

linear_reg = LinearRegression()

linear_reg.fit(X_train, y_train)

#R2 scores
print (f'Linear Regression Training R2 Score: {round(linear_reg.score(X_train, y_train), 4)}.')
print (f'Linear Regression Testing R2 Score: {round(linear_reg.score(X_test, y_test), 4)}.')

Unfortunately, even with many new features, the we are not having much success with regression to predict the target. In the next notebook we'll try a few more regression methods
