==> code/06_Clustering.ipynb <==
# Check (That Tweet) Yo Self 
## Prioritizing Tweets to Fact Check
###### Part 6: Clustering (Unsupervised Learning)
In the previous notebooks, we collected tweets about Coronavirus from around the time of Donald Trump's lysol comment and connected them with a separate pull on user data.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.linear_model import LinearRegression, LogisticRegression
from sklearn import metrics
from sklearn.ensemble import RandomForestRegressor, ExtraTreesRegressor, AdaBoostRegressor
from sklearn.tree import DecisionTreeRegressor
from sklearn.svm import LinearSVC
from sklearn.model_selection import GridSearchCV
import statsmodels.api as sm
from sklearn.preprocessing import StandardScaler, PolynomialFeatures
from sklearn.ensemble import BaggingRegressor
from sklearn.decomposition import PCA
from sklearn.cluster import KMeans, DBSCAN
from sklearn.metrics import silhouette_score
from nltk.sentiment.vader import SentimentIntensityAnalyzer
import pickle
from sklearn.neighbors import KNeighborsClassifier

Now that we have gathered all the neccesary user information, we are going to cluster based on user stats.

tweet = pd.read_csv('../data/tweet_users.csv')

tweet = tweet[:33200]

First, we'll engineer a few features that takes a closer look at the tweet text and user info. Descriptions for each features are below:
- **'len_user'** : lenth of the username
- **'big_feelings'** : percentage of the tweet in uppercase
- **'ratio'** : user's followers divded by how many they are following
- **'has_url'** : user has a url as part of their profile
- **'has_location'** : user has a location listed on their profile
- **'has_bio'** : user has a bio as part of their profile
- **'len_bio'** : length of user's bio
- **'ratio_num_user'** : percentage of numeric characters in the username
- **'emotional_range'** : idea stemmed from this [paper](https://hpi.de/fileadmin/user_upload/fachgebiete/naumann/publications/2013/Analyzing_and_Predicting_Viral_Tweets.pdf), it's the absoluate value of postive and negative emotion added together and divided by 10 to explain the range of emotions used in the tweet

tweet['len_user'] = [len(x) for x in tweet['author']]

def per_upper(string):
    count = 0
    for s in string:
         if s == s.upper():
                count += 1
    ratio = count / len(string)
    return ratio

tweet['big_feelings'] = tweet['text'].apply(per_upper)

def get_ratio(followers, following):
    if following == 0:
        following = 1
    elif followers == 0:
        return 0
    else:
        return followers / following

tweet['ratio'] = [get_ratio(m, n) for m, n in zip(tweet['user_followers'], tweet['user_following'])]

tweet['has_url'] = tweet['user_url'].notna().astype(int)
tweet['has_location'] = tweet['user_location'].notna().astype(int)
tweet['has_bio'] = tweet['user_bio'].notna().astype(int)

tweet['len_bio'] = [len(str(x)) for x in tweet['user_bio']]

def is_numeric(string):
    count = 0
    for s in string:
        try:
            int(s)
            count += 1
        except:
            count = count
    return count / len(string)

tweet['ratio_num_user'] = tweet['author'].apply(is_numeric)

sent = SentimentIntensityAnalyzer()

def emotion_range(string):
    emo_r = np.abs(sent.polarity_scores(string)['neg']) + sent.polarity_scores(string)['pos']
    return emo_r / 10

tweet['emotional_range'] = tweet['text'].apply(emotion_range)

tweet.drop(columns = ['user_bio', 'user_location', 'user_url'], inplace = True)

From earlier investigation of specific accounts, we concluded if a person has "null" favorites, it is because they have not favorited anything so we can impute "0" for the null values here.

tweet['user_favorites'] = tweet['user_favorites'].fillna(0)

Likewise for ratio, the values showing up as null are when they have 0 followers and 0 following. While this might be a ratio of 1 (equal followers/following), this really shows a lack of connection and engagement so it would be more appropriate to fill in 0.

tweet['ratio'] = tweet['ratio'].fillna(0)

tweet.dropna(inplace = True)

tweet.info()

We now have no missing values and can continue to clustering.

These are the user attributes we want to focus our clustering on:

user_cat = ['user_tweets', 'user_following', 'user_followers', 
             'ratio', 'has_url', 'has_location', 'has_bio']

to_cluster = tweet[user_cat]

Scaling the data before creating clusters:

ss = StandardScaler()
ss.fit(to_cluster)

z_cluster = ss.transform(to_cluster)

After try a range of different values for number of clusters, 5 gave the highest jump in silhouette score while still having distinct characteristics between the categories.

km = KMeans(n_clusters=10, random_state = 21)
km.fit(z_cluster)

db = DBSCAN(eps = .8, min_samples = 1000)

db.fit(z_cluster)

silhouette_score(z_cluster, km.labels_)

silhouette_score(z_cluster, db.labels_)

Our DBSCAN score is slightly higher and is producing more logical groupings.

db.labels_

Pickling this model to use later on:

pickle.dump(db, open('../models/dbscan.pkl', 'wb'))

pickle.dump(ss, open('../models/standardscaler.pkl', 'wb'))

tweet['user_group_db'] = db.labels_

Checking out the distribution of our clusters:

tweet['user_group_db'].value_counts()

Adding these cluster groups to the DataFrame.

Even though our clusters didn't take into account the specific "reach" (reply, favorite, and retweet count) for the tweet, we want to see how different clusters measure up in this capacity.

tweet['target'] = tweet['reply_count'] + tweet['favorite_count'] + tweet['retweet_count']

tweet.head(2)

Saving this DataFrame with new features and cluster assignments to a csv for further EDA in the next notebook.

tweet.to_csv('../data/user_cluster_tweets.csv', index = False)

tweet.shape

We realized later on that DBSCAN doesn't make prediction so we would have to train a classifer on the labels in order to analyze new information:

knn = KNeighborsClassifier()

X_knn = z_cluster
y_knn = tweet['user_group_db']

Xk_train, Xk_test, yk_train, yk_test = train_test_split(X_knn, y_knn, random_state = 21)

knn.fit(Xk_train, yk_train)

knn.score(Xk_train, yk_train)

knn.score(Xk_test, yk_test)

Pickling this model for use later on

pickle.dump(knn, open('../models/knn.pkl', 'wb'))
