==> code/09_Cluster_NLP_EDA.ipynb <==
# Check (That Tweet) Yo Self 
## Prioritizing Tweets to Fact Check
###### Part 9: Cluster NLP EDA
This notebook looks at the top words and phrases for each cluster

Import necessary libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import requests
import time
import warnings
import regex as re
import seaborn as sns
import re
import statistics

from sklearn.pipeline import Pipeline, FeatureUnion
from sklearn.preprocessing import StandardScaler, FunctionTransformer
from sklearn.linear_model import LogisticRegression, LinearRegression
from sklearn.linear_model import Ridge 
from sklearn.linear_model import Lasso
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.metrics import confusion_matrix, accuracy_score, classification_report

from nltk.sentiment.vader import SentimentIntensityAnalyzer
warnings.filterwarnings('ignore')
np.random.seed(824)
from bs4 import BeautifulSoup 

# Import stopwords.
from nltk.corpus import stopwords # Import the stopword list
import nltk

from tweetscrape.users_scrape import TweetScrapperUser

%config InlineBackend.figure_format = 'retina'
%matplotlib inline

plt.style.use('fivethirtyeight')

Read in csv

tweet = pd.read_csv('../data/user_cluster_tweets.csv')

tweet.head()

Additional cleaning to the clean tweets column with custom stopwords that are significant in number but hold no contextual value

# sklearn's stopwords, extracted
sklearn_stopwords = list(CountVectorizer(stop_words = 'english').get_stop_words())
#Custom created list
custom_stopwords = ['com',
                    'twitter',
                    'pic',
                    'http',
                    'isolation',
                    'pandemic',
                    'covid',
                    'quarantine',
                    'vaccine',
                    'coronavirus',
                    'lysol',
                    'ingest',
                    'inject',
                    'disinfectant',
                    'bleach',
                    'don',
                    've']
# Personalized stopwords
personal_stopwords = sklearn_stopwords + custom_stopwords

def tweet_to_words(raw_tweet):
    
    # 1. Remove HTML.
    tweet_text = BeautifulSoup(raw_tweet).get_text()
    
    # 2. Remove non-letters.
    letters_only = re.sub("[^a-zA-Z]", " ", tweet_text)
    
    # 3. Convert to lower case, split into individual words.
    words = letters_only.lower().split()
    
    # 4. Convert personalized stopwords to set
    stops = set(personal_stopwords)

    # 5. Remove stopwords.
    meaningful_words = [w for w in words if w not in stops]
    
    # 6. Join the words back into one string separated by space, 
    # and return the result.
    return(" ".join(meaningful_words))

# Get the number of tweets based on the dataframe size.
total_tweets = tweet.shape[0]
print(f'There are {total_tweets} tweets related to coronavirus.')

# Initialize an empty list to hold the clean tweets.
clean_tweets = []

print("Cleaning and parsing twitter data...")

# Instantiate counter.
j = 0

# For every post in our training set...
for string in tweet['text_links_removed']:
    
    # Convert post to words, then append to clean_train_posts.
    clean_tweets.append(tweet_to_words(string))
    
    # If the index is divisible by 1000, print a message.
    if (j + 1) % 5000 == 0:
        print(f'Tweet {j + 1} of {total_tweets}.')
    
    j += 1 

tweet = tweet.assign(clean_text = clean_tweets)

tweet.groupby(['user_group_db']).mean().sort_values(['target'])

db_outlier = tweet[tweet['user_group_db'] == -1]
db_0 = tweet[tweet['user_group_db'] == 0]
db_1 = tweet[tweet['user_group_db'] == 1]
db_2 = tweet[tweet['user_group_db'] == 2]
db_3 = tweet[tweet['user_group_db'] == 3]
db_4 = tweet[tweet['user_group_db'] == 4]

def top_trends(df, column, ngram_min, ngram_max, top_word_count):
    cvec = CountVectorizer(stop_words = 'english', min_df=1, max_df=0.25, ngram_range=(ngram_min, ngram_max))
    term_mat = cvec.fit_transform(df[column])
    print(f'Number of unique items: {len(cvec.get_feature_names())}')
    print()
    term_df = pd.DataFrame(term_mat.toarray(), columns=cvec.get_feature_names())
    top_words = pd.DataFrame(term_df.mean().sort_values(ascending = False).head(top_word_count), columns = ['Frequency'])
    return top_words

### Grabbing Phrases for Entire Dataframe

tweet_one = top_trends(tweet, 'clean_text', 1, 1, 20)
tweet_one.to_csv('../data/tweet_one.csv')

tweet_two = top_trends(tweet, 'clean_text', 2, 2, 20)
tweet_two.to_csv('../data/tweet_two.csv')

tweet_three = top_trends(tweet, 'clean_text', 3, 3, 20)
tweet_three.to_csv('../data/tweet_three.csv')

tweet_four = top_trends(tweet, 'clean_text', 4, 4, 20)
tweet_four.to_csv('../data/tweet_four.csv')

### Phrases for Outlier Group

outlier_one = top_trends(db_outlier, 'clean_text', 1, 1, 20)
outlier_one.to_csv('../data/outlier_one.csv')

outlier_two = top_trends(db_outlier, 'clean_text', 2, 2, 20)
outlier_two.to_csv('../data/outlier_two.csv')

outlier_three = top_trends(db_outlier, 'clean_text', 3, 3, 20)
outlier_three.to_csv('../data/outlier_three.csv')

outlier_four = top_trends(db_outlier, 'clean_text', 4, 4, 20)
outlier_four.to_csv('../data/outlier_four.csv')

### Phrases for our main priority to focus in on when identifying fake news tweets (group 3)

concern_one = top_trends(db_3, 'clean_text', 1, 1, 20)
concern_one.to_csv('../data/concern_one.csv')

concern_two = top_trends(db_3, 'clean_text', 2, 2, 20)
concern_two.to_csv('../data/concern_two.csv')

concern_three = top_trends(db_3, 'clean_text', 3, 3, 20)
concern_three.to_csv('../data/concern_three.csv')

concern_four = top_trends(db_3, 'clean_text', 4, 4, 20)
concern_four.to_csv('../data/concern_four.csv')

tweet.groupby(['user_group_db']).mean().sort_values(['target'])

### 2nd priority concern (group 0)

priority_2_one = top_trends(db_0, 'clean_text', 1, 1, 5)
priority_2_one.to_csv('../data/priority_2_one.csv')

priority_2_four = top_trends(db_0, 'clean_text', 4, 4, 5)
priority_2_four.to_csv('../data/priority_2_four.csv')

### 3rd priority concern (group 4)

priority_3_one = top_trends(db_4, 'clean_text', 1, 1, 5)
priority_3_one.to_csv('../data/priority_3_one.csv')

priority_3_four = top_trends(db_4, 'clean_text', 4, 4, 5)
priority_3_four.to_csv('../data/priority_3_four.csv')

### 4th priority concern (group 2)

priority_4_one = top_trends(db_2, 'clean_text', 1, 1, 5)
priority_4_one.to_csv('../data/priority_4_one.csv')

priority_4_four = top_trends(db_2, 'clean_text', 4, 4, 5)
priority_4_four.to_csv('../data/priority_4_four.csv')

### 5th priority concern (group 1)

priority_5_one = top_trends(db_1, 'clean_text', 1, 1, 5)
priority_5_one.to_csv('../data/priority_5_one.csv')

priority_5_four = top_trends(db_1, 'clean_text', 4, 4, 5)
priority_5_four.to_csv('../data/priority_5_four.csv')

# Round em up!
Putting all the phrases for each group back into one DataFrames for each phrase length.

outlier_one = pd.read_csv('../data/outlier_one.csv')
outlier_four = pd.read_csv('../data/outlier_four.csv')
concern_one = pd.read_csv('../data/concern_one.csv')
concern_four = pd.read_csv('../data/concern_four.csv')
priority_2_one = pd.read_csv('../data/priority_2_one.csv')
priority_2_four = pd.read_csv('../data/priority_2_four.csv')
priority_3_one = pd.read_csv('../data/priority_3_one.csv')
priority_3_four = pd.read_csv('../data/priority_3_four.csv')
priority_4_one = pd.read_csv('../data/priority_4_one.csv')
priority_4_four = pd.read_csv('../data/priority_4_four.csv')
priority_5_one = pd.read_csv('../data/priority_5_one.csv')
priority_5_four = pd.read_csv('../data/priority_5_four.csv')

all_tweets_one = pd.read_csv('../data/tweet_one.csv')
all_tweets_four = pd.read_csv('../data/tweet_four.csv')

outlier_one = outlier_one.head()
outlier_four = outlier_four.head()
concern_one = concern_one.head()
concern_four = concern_four.head()

outlier_one = outlier_one.rename(columns = { 'Unnamed: 0' : 'Outlier'})        
concern_one = concern_one.rename(columns = { 'Unnamed: 0' : '1st Priority'})        
priority_2_one = priority_2_one.rename(columns = { 'Unnamed: 0' : '2nd Priority'}) 
priority_3_one = priority_3_one.rename(columns = { 'Unnamed: 0' : '3rd Priority'}) 
priority_4_one = priority_4_one.rename(columns = { 'Unnamed: 0' : '4th Priority'}) 
priority_5_one = priority_5_one.rename(columns = { 'Unnamed: 0' : '5th Priority'}) 


outlier_four = outlier_four.rename(columns = { 'Unnamed: 0' : 'Outlier'})             
concern_four = concern_four.rename(columns = { 'Unnamed: 0' : '1st Priority'})            
priority_2_four = priority_2_four.rename(columns = { 'Unnamed: 0' : '2nd Priority'})
priority_3_four = priority_3_four.rename(columns = { 'Unnamed: 0' : '3rd Priority'}) 
priority_4_four = priority_4_four.rename(columns = { 'Unnamed: 0' : '4th Priority'})
priority_5_four = priority_5_four.rename(columns = { 'Unnamed: 0' : '5th Priority'}) 

outlier_one

top_one_word = outlier_one[['Outlier']]
top_one_word['1st Priority'] = concern_one['1st Priority']
top_one_word['2nd Priority'] = priority_2_one['2nd Priority']
top_one_word['3rd Priority'] = priority_3_one['3rd Priority']
top_one_word['4th Priority'] = priority_4_one['4th Priority']
top_one_word['5th Priority'] = priority_5_one['5th Priority']

top_one_word.to_csv('../data/clusters_top_one_word.csv', index = False)

top_one_word

Above are the top words for each cluster.

top_four_word = outlier_four[['Outlier']]
top_four_word['1st Priority'] = concern_four['1st Priority']
top_four_word['2nd Priority'] = priority_2_four['2nd Priority']
top_four_word['3rd Priority'] = priority_3_four['3rd Priority']
top_four_word['4th Priority'] = priority_4_four['4th Priority']
top_four_word['5th Priority'] = priority_5_four['5th Priority']

top_four_word.to_csv('../data/clusters_top_four_word.csv', index = False)

top_four_word

These are the top 4 word phrases for each cluster.

all_tweets_one = pd.read_csv('../data/tweet_one.csv')

all_tweets_one = all_tweets_one.head(10)

all_tweets_four = all_tweets_four.head(10)

all_tweets_one = all_tweets_one.rename(columns = { 'Unnamed: 0' : 'Top Words'})

all_tweets_four = all_tweets_four.rename(columns = { 'Unnamed: 0' : 'Top Groups of Words'})

top_tweet_stats = all_tweets_one[['Top Words']]
top_tweet_stats['Top Groups of Words'] = all_tweets_four['Top Groups of Words']

top_tweet_stats

These are the top overall words and phrases for all tweets.
